fix: move engine_str to config.ssot to respect import layers
config.models imported engine_str from domain.experiment, breaking the
config-must-not-import-domain contract. engine_str operates on the Engine
enum (which lives in config.ssot), so it belongs there; study importers
repointed accordingly.
